Mudanca de senha wifi e tipo da variavel ano

+++ WiFiServer/ServerSocet.ino
@@ -118,7 +118,7 @@ int treatmentRequest(JsonObject& noSensor){
     Serial.println("Begin to format");
     Serial.println(dateTime);
    
-    noSensor["ano"]= getDate('-').toInt();
+    noSensor["ano"]= getDate('-');
     noSensor["mes"]= getDate('-');
     noSensor["dia"]= getDate('-');
     noSensor["hora"]= getDate(':').toInt();
